Moved the actual dotfiles into their own directory.
I want to rewrite the install script to use bash (so that I can run it
on a machine that doesn't have ruby installed). Moving the dotfiles into
their own directory allows me to remove a bit of logic from the install
script, which I'm hoping makes it slightly easier to translate to bash.